@@ -1,7 +1,7 @@
 module ApplicationHelper
 
   def flash_class(level)
-    case level
+    case level.to_sym
     when :notice then 'info'
     when :error then 'danger'
     when :alert then 'warning'
